requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aioresponses==0.7.6
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
astroid==2.9.3
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
black==24.4.2
blinker==1.8.2
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@2a420edb307c9bdf640f036d3b196bed474b8593
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
grpcio==1.65.1
hydra-core==1.3.2
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
inference-sdk==0.15.1
iopath==0.1.9
isort==5.10.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy-object-proxy==1.7.1
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
mccabe==0.6.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
opencv-contrib-python==4.5.5.62
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==2.5.1
portalocker==2.10.1
protobuf==4.25.4
py-cpuinfo==9.0.0
pycocotools==2.0.8
pylint==2.12.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.37
scipy==1.13.1
shapely==2.0.5
six==1.16.0
supervision==0.22.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
toml==0.10.2
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
UNKNOWN @ git+https://github.com/THU-MIG/yolov10.git@cd2f79c70299c9041fb6d19617ef1296f47575b1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.13.3
yacs==0.1.8
yarl==1.9.4
zipp==3.19.2
